[Fix] 전체분석 페이지 코드 수정
- 작업 내용 -
1. 전체분석 페이지 폴더 생성 후 이동
2. %에 맞게 자동으로 여유, 적절, 도전, 위험으로 분류되어 화면에 보이도록 코드 수정
