+++ src/index.js
@@ -7,10 +7,8 @@ import "assets/vendor/font-awesome/css/font-awesome.min.css";
 import "assets/scss/argon-design-system-react.scss?v1.1.0";
 
 import Index from "views/Index.js";
-import FullAnalysis from "views/examples/analysis/FullAnalysis.js";
+import FullAnalysis from "views/examples/analysis/Full/FullAnalysis.js";
 import IntensiveAnalysis from "views/examples/analysis/IntensiveAnalysis/IntensiveAnalysis.js";
-import MapAnalysis from "views/examples/analysis/MapAnalysis.js";
-import IntensiveAnalysis from "views/examples/analysis/IntensiveAnalysis.js";
 import MapAnalysis from "views/examples/analysis/Map/MapAnalysis.js";
 import Landing from "views/examples/Landing.js";
 import Login from "views/examples/Login.js";
