+++ src/views/examples/analysis/Full/FullAnalysis.js
